Binary Clock written in x86 Assembly
I wish I commented this code… It was written from scratch for an
independent study in school as a final project. It converts the system
time into a readable format and displays it on the screen. Above each
digit in the time is a visual binary representation of the digit.